Use deployed helper SC to estimate account balances (#3551)
# Description
This is a follow-up to #3550, which uses the same approach for fetching
account balances. More details about the approach can be found in the
corresponding PR.

## How to test
Existing tests. Also, it was tested on Lens in #3545(upd and on mainnet
prod).

The CoW AMM forked test can't use a new block since this leads to
significant indexing delays and test failures, so the Balances SC gets
deployed manually right in the test.