add missing `creation_timestamp` column (#3534)
# Description
added to tables: `app_data` and `order_quotes`. This makes them aligned
with the other table schemas and easier for our analytics team to
process incrementally instead of having to reindex entirely.

Find more context
[here](https://cowservices.slack.com/archives/C036PMLUPQF/p1753890884620899)

## Note
Not 100% sure about the ethflow query change, please double check if it
was the originals query intent to use the `order` timestamp here?

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>