[Driver] expose endpoint to get current effective gas price (1/3) (#3567)
# Description
We are currently using different logic when estimating the gas price in
the API (for network cost estimation purposes) compared to when
estimating it in the driver (for actual submission purposes).

The difference in resulting gas price can be significant (up to 3-5x on
mainnet). This may lead to us underreporting the network cost during
quoting. Especially for small orders, where the network cost is a large
relative fraction of the total volume, the error can easily be larger
than the user's slippage tolerance leading to the order not being
matched.

As a solution I suggest exposing the gas price computation that the
driver performs via an API and calling this API from within the api (and
autopilot) during quoting.

# Changes
This PR is the first of 3, it exposes a new `/gasprice` endpoint on the
top level driver API (not on each solver's sub-api given that the gas
price computation isn't solver specific).

## How to test
On a locally running driver:

```
curl http://localhost:11088/gasprice 
```