Don't store liquidity tick net (#3579)
# Description
This is a some kind of follow-up to #3578, where I noticed that
liquidity tick net is not used when working with concentrated univ3
liquidity sources in the Baseline solver, while this data might be heavy
and contain thousands of records for each pool, which can potentially
lead to unreasonable memory consumption.

# Changes
Do not store the liquidity tick net received from driver. It would be
great to not even parse it, but a shared struct is being used here, so
the parsed auction is now consumed by the `into_domain` function to be
dropped as soon as possible to avoid storing it in memory for the whole
`solve` operation.

## How to test
No tests so far for univ3 liquidity due to annoying issues with its
fetching delays in e2e tests.