Make high_gas_limit work with new foundry version (#3586)
# Description

We fixed foundry version for driver tests in
https://github.com/cowprotocol/services/pull/3561 because high_gas_limit
started failing. This PR fixes the test with the latest foundy version
and adds comments that make it clear what we're testing. I also added a
test that makes sure we discard solutions that take more gas than half
of the block gas limit.