Dynamically adjust timeout in native price requests (#3585)
# Description
For an order to be put into the current auction the autopilot needs to
know the native price of the sell and the buy token. Any delays in
fetching the native prices can result in orders having "to sit out" an
auction which immediately adds at least 12s to the execution time. But
we also have to be mindful of the time we spend fetching native prices
because that delays the entire auction.

Also in order to reduce cost we fetch native prices in stages (first the
cheap quoters, then the expensive ones).

# Changes
So far we evenly divided the time for each stage due to its simplicity.
However, we are currently dealing with issues where the later stages
just need a little bit more time. To get the most out of the allocated
time and use it fully this PR adjusts the way we compute the timeout for
the current stage.
Now the timeout of the current stage will adjust dynamically. The
current stage always gets `total_time_left / stages_left`. This does not
change anything for the very first stage but subsequent stages would get
more time whenever some stage does not use the entire allocated time
budget.
For the native price estimation code to know how many stages are left we
build a new `Context` struct in the `CompetitionEstimator` that contains
that information.

I also `async-ified` the other tests which turned out to be unnecessary.
I think the new way is cleaner but LMK if you prefer to have that in a
separate PR.

## How to test
Added a unit test to assert that the timeouts are what we'd expect them
to be (left code comments to explain the subtle details)