Init alloy migration PoC (#3523)
# Description
Just a quick draft to briefly explain that migration plan for alloy.
This is best understood by reading the few commits individually.

# Changes
1. introduce a wrapper type that contains the old web3 and the new alloy
provider. To make it most ergonomic with our current code it derefs into
the web3 provider. With that available everywhere we can easily start
replacing individual uses of the old stuff with alloy.
2. implement conversion types (just some helper traits to make
converting between web3 and alloy types painless)
3. changes to generate bindings using alloy. Uses a custom macro to
allow us to attach deployment info with each contract like we do today.
This should probably be rethought but for the sake of fewer breaking
changes I decided to reimplement the current approach.
4. example commit swapping out bindings from `ethcontract-rs` with alloy
bindings