serialize with alloy's number type (#3618)
# Description
With all the recent driver performance improvements the most amount of
CPU time is used for serializing the requests that get forwarded to the
connected solver engines. And the majority of that time is used for
printing `U256` as a decimal string.

# Changes
It turns out that `primitive_types::U256::to_string()` is so slow that
it's still considerably faster to first convert it to `alloy`'s `U256`
type and serializing that instead.

## How to test
ran a quick test on shadow mainnet to collect flamegraphs. With this
change serializing the request goes from ~35% of the total CPU time to
~25%.

Before
<img width="1728" height="850" alt="Screenshot 2025-09-01 at 20 54 46"
src="https://github.com/user-attachments/assets/b273faca-203b-402c-bcb3-5fd18aae7b01"
/>

After
<img width="1727" height="834" alt="Screenshot 2025-09-01 at 20 55 01"
src="https://github.com/user-attachments/assets/c1eb5cd3-e086-42aa-bdef-c4b839fc1e45"
/>