Alloy Account type conversion (#3599)
# Description
Adds a conversion from `ethcontract::Account` to the corresponding alloy
types. A wrapper is introduced because local accounts don’t share a
common type with KMS or PrivateKey in alloy. The usage also differs when
building/signing transactions: for a local account it’s enough to just
set the from field (an address), while with KMS and PrivateKey there
isn’t a direct way to use a signer that’s different from the provider’s
configured one(more details
https://github.com/alloy-rs/alloy/issues/2829).

`ethcontract::Account::Locked` remains unused in this repo and it seems
like there is no direct way to convert it into
`alloy::signers::local::LocalSigner`, so this branch is `unimplemented`.

This PR also depends on
https://github.com/cowprotocol/ethcontract-rs/pull/981, which explains
incompatibilities.

After ethcontract was updated in
https://github.com/cowprotocol/ethcontract-rs/pull/981, driver tests
started failing with stack overflow. As @jmg-duarte's research showed,
that the issue lies in how updated third-party libraries optimizations
work, which is not related to any infinite recursion. The problem
doesn't exist when building with the `--release` flag, so instead, the
stack size was increased.

## How to test
In the upcoming PRs.

---------

Co-authored-by: MartinquaXD <martin.beckmann@protonmail.com>