Add deployment block info to alloy bindings (#3630)
# Description
To continue with the alloy migration, a deployed block info is required
for some of the SC, list GPV2Settlement, BalancerV2.x, others.

# Changes
Extend the bindings macro and InstanceExt interface with that data. Now,
in the bindings! macro, added a way to provide either only address or a
tuple with (address, deployment block).