Optimize invalid signature filtering (#3647)
# Description
Checking signatures is currently one of the slower auction building
steps. This PR implements a couple optimizations to speed things up.

# Changes
- avoid `Itertools::partition` as that moves around every item even if
we only need to move a few (e.g. in this case we don't need to split ALL
orders into pre-signed and "unknown" - instead we just work with the
orders that are actually relevant for this processing step)
- changed `validate_signatures()` to only handle a single check instead
of multiple - that made it a lot less awkward to figure out for which
order the check failed.
- updated internal helper functions to take an owned `SignatureCheck` so
we don't have to clone the `signature` and `interaction` bytes twice
- used `FuturesUnordered` instead of `join_all` to avoid allocating a
vector for all the results - instead we just await each future
concurrently and push the invalid orders into the result vector

## How to test
updated the existing unit test
unfortunate the performance improvement of this PR can only really be
measured by deploying to mainnet prod as all the staging environments
don't have enough data to matter and the shadow environment does not hit
this code path.