Opt-out of quote verification (#3650)
# Description
This PR is really stupid but unfortunately necessary at the moment...
Some tokens are weird to route optimally and only few solvers support
them. However, anybody can create a uniswap pool for these tokens which
may cause any good solver to be able to route these tokens
automatically.
This is generally a good thing except in one particular edge case:
If the "good" solver is not able to provide calldata with their quote
that can be verified but the "bad" solver's calldata verifies just fine
we'll currently show the quote from the bad solver despite having very
high confidence that the other solver is actually better and knows how
to route the tokens.

To avoid showing terrible prices in this VERY specific scenario we need
an escape hatch to conditionally bypass the quote verification. I
decided to go for a simple token list. If either the sell or the buy
token show up in the list that we define we'll skip the quote
verification altogether.
That way quote from all solvers (good and bad) will be unverified which
levels the playing field and we again pick the quote based on the
highest out amount.

---------

Co-authored-by: ilya <ilya@cow.fi>