Add MutWallet to the testing provider (#3651)
Adds the MutWallet to enable adding signers when needed.

This is a "solution" to alloy-rs/alloy#2829
Since Alloy's Provider is built on layers and in the end it seals them up, it becomes impossible to change the signing wallet. The solution is to share the wallet behind the scenes through the classic Arc + lock (Rw in this case)

The main challenge here was running tokio's locks in "sync" contexts:

    During tests we have "current_thread" so we can't run blocking operations (.blocking_read), we can however, run a separate thread and wait for it (ironic because it blocks the thread anyway — though this is ok because it's for tests)
    Outside tests, we're running the multi-thread runtime, thus spawning an extra thread and waiting is a waste because we're blocked waiting for it anyways, using block_in_place we can signal the runtime that we're in a pickle and we're going to block, evicting other tasks and keeping the system chugging along