Generalize the event handling module (#3648)
# Description
This is the first prerequisite for the upcoming PRs that generalize the
event handling module, making it possible to integrate with Alloy smart
contract instances. Instead of relying on `ethcontract-rs`–specific
types in the interface, the trait is now generic and can work with
primitives from either `ethcontract-rs` or `alloy`.

All the interfaces in the `event_handling.rs` file don't depend on the
`ethercontract-rs` crate, meaning we can use any underlying
implementation in the future.

Some common logic was extracted to the `EthcontractEventRetrieving`
trait. This is required to remove `ethcontract-rs` crate dependency from
the `EventRetrieving` trait and deduplicate the code.

The change looks pretty straightforward, but you can't imagine how many
failed attempts were made to reach its current state 🤕