Add BalancerQueries contract (#3646)
# Description
This PR adds a helper contract called BalancerQueries to the contracts
crate.
Giving a bit more context, currently the Balancer solver relies on SOR
quotes only, which might be running on slightly out-of-date data from
the Balancer API.
Being able to update and validate amounts with an extra on-chain query
swap step would be ideal.
On Balancer V3 it's possible to query swaps with the BalancerRouter,
which is already present in the contracts crate, but on Balancer V2 the
best way is to use this BalancerQueries helper contract, that is not yet
available.

# Changes
I simply followed the steps to add a new contract provided in the
[contracts readme.md
file](https://github.com/cowprotocol/services/blob/main/crates/contracts/README.md)
> - In vendor.rs extend ARTIFACTS with the package and contract name.
> - Run the vendor binary for example with (cd crates/contracts; cargo
run --bin vendor --features bin). This creates a new json file for this
contract in the artifacts folder.
> - In build.rs add a generate_contract call for the contract. This
creates the ethcontract generated rust code file.
> - In lib.rs add an include! call for the contract. This imports the
rust code into the library.

## How to test
Not sure this addition requires tests to be added. If you think it's
necessary, please share a template or even some example I can follow,
because I couldn't find test examples from the existing contracts.

## Related Issues
Gettin this PR merged and released will unblock
https://github.com/gnosis/solvers/pull/161

---------

Co-authored-by: José Duarte <duarte.gmj@gmail.com>