Retry mechanism for settlement indexing (#3661)
# Description
Recently Haris reported that a few settlements were not fully indexed on
base. The investigation showed that there were temporary nodes issues
that caused the RPC requests to time out.
Due to the lack of a retry mechanism we aborted the indexing process of
these auctions fully.

# Changes
- moved logic to aggregate the necessary settlement data into a separate
function
- replaced `Result<bool>` with `Result<IndexingSuccess>` to make it
clearer in the caller what happened
- propagated all temporary errors (DB or RPC issues) as retryable errors

## How to test
not clear - the error cases are very tedious to mock