fake flashloans in balance checks (#3674)
# Description
This is the first PR of many to gradually make user balance checks and
order signature checks compatible with flashloans.
For context when a user wants to place an order we make sure that their
account has sufficient allowance and balance for the sell token. When
placing a flashloan order it may very well be the case that the user has
0 allowance or balance but their order includes a pre-hook that sets up
the balance and allowance correctly.
Without this PR the balance simulation logic was completely unaware of
the flashloan requirement so placing such orders would be rejected due
to missing balances and allowances.

To make the balance simulation logic compatible with flashloans we
adjust it to fake the effects of a flashloan. Specifically we compute
the exact state override that is needed to change the flashloan token
such that the flashloan receiver has the required balance.

# Changes
- adjusted `BalanceOverriding` interface slightly to make it easier to
work with
- add `Arc<dyn BalanceOverriding>` and use it to set the required state
overrides in the balance simulation logic
- autopilot and orderbook create an instance based on CLI arguments
- driver creates instance with sensible defaults to make things simpler
for now
- removed some balance related dead code from the `Erc20` struct

As a follow up we could still optimize how the individual components
setup and share the `BalanceOverrides` instance. But that only becomes
relevant with tons of flashloan orders for many different tokens.

## How to test
all tests should still pass (we just replaced a less versatile hack with
a reasonable solution that should work for all reasonable flashloan
setups)