Add CLI flag for filtering orders in autopilot (#3688)
# Description
Currently the autopilot filters out tons of orders from the current
auction as an optimization.
In order to be less opinionated on what orders can be settled at any
point in time these checks should not happen on a protocol level (i.e.
in the autopilot). However, since things like balance checking make
sense in the grand scheme of things these checks how happen in the
reference driver.

So effectively disabling order filtering on the autopilot side should
only have an impact on colocated drivers that are connected to the
autopilot.
Since it's not trivial to see what would happen if this check were to be
removed right away I decided to feature gate the logic. That way we can
test temporarily on prod to see if anything bad happens and revert
quickly if it does.

# Changes
- introduce a flag that makes order filtering optional