Notify solvers on serialization error (#3682)
# Description
This PR adds a new notification to the reference driver to notify
solvers when we are unable to parse a solver's response to the `/solve`
request.

# Changes
- [ ] Add SerializationError to notify
- [ ] Send SerializationError when we fail to parse a solver's response

## How to test
<!--- Include details of how to test your changes, including any
pre-requisites. If no unit tests are included, please explain why and
how to test manually
1.
2.
3.
-->

## Context
Currently a large amount of time for new solvers in the shadow
competition is spent debugging this issue. If we notify solvers on a
serialization error this will allow solvers to debug this issue
independently.

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>