Reuse database connection if db_read_url is not provided (#3723)
# Description
The recent changes to autopilot and orderbook allow to specify
DB_READ_URL and DB_WRITE_URL. The former is used only for non mutating
database operations. The plan is to point DB_READ_URL to our
read-replica (giving the possibility of slightly outdated data). The
DB_WRITE_URL points to the main database.

# Changes
Added reuse of connection pool in case of DB_READ_URL and DB_WRITE_URL
having the same values.

## How to test
1. Run E2E tests
2. Check the deployment on staging

---------

Co-authored-by: ilya <ilya@cow.fi>