Drop V092 Flyway migration (#3751)
In #3747, the `auction_orders` table is dropped, but since we now run 2
autopilots in parallel on each service restart, dropping the table right
away would cause panics in the currently active autopilot, which is not
aware of the latest changes. This PR drops the migration that would need
to be added only in the next release.

Since the previous PR was already merged, it has been applied on staging
already, which now requires manually altering the
`flyway_schema_history` on each chain before merging this PR 😓