Migrate from `auction_participans` to `proposed_solutions` table (#3766)
# Description
The `auction_participans` table has become obsolete, since the same data
is now stored in the `proposed_solutions`. This PR migrates usage to the
latter table. The migration script will be a part of #3753, since this
is a breaking change, which requires gradual deployment to avoid panics
when running 2 auctions in parallel.