Fix refunder wallet (#3779)
# Description
After the CoWSwapEthFlow SC was migrated to alloy, we've started
receiving the following errors:
```
2025-10-14T16:46:08.171Z  WARN refunder: Error while refunding ethflow orders: local usage error: Missing signing credential for 0x0214aE5fD178986fA18ff792e0b995Dc6a78cD56  
Caused by:     Missing signing credential for 0x0214aE5fD178986fA18ff792e0b995Dc6a78cD56
```
That basically means that our refunder doesn't work.

The reason for that is that the refunder's private key wasn't set
properly in the alloy provider. This PR fixes it.

## How to test
Probably not easy.