Add tx gas limit to driver config (#3783)
# Description
Required for tx gas configuration to be backwards compatible and
guarantee smooth deployment.

The infrastructure will be able to set tx gas limit per chain, but
unless this field is expected, the driver would just panic on parsing
the config.

This introduces the required field as an option, and does not do
anything with the value if set.
https://github.com/cowprotocol/services/pull/3780 will actually make it
mandatory and work.