Migrate BalancerV3BatchRouter to alloy (#3775)
# Description
Migrates BalancerV3BatchRouter to alloy.

Alloy, by default, wasn't able to parse the ABI JSON we currently have,
so I had to update it a bit. To see the exact change, see
[this](https://github.com/cowprotocol/services/pull/3775/commits/d4db354ffb7c2cf33b1cfd7aa0cd17856a06bf5c#diff-6b43d7a940f1bbb3e887415e980543e0825ea48d1d5919797012f188f71b8551)
commit after the formatting one.

This would require changes in the gnosis/solvers repo, since the SC is
used only there.