[TRIVIAL] pin postgres version to 16 (#3814)
# Description
Our `posgres` version is currently not pinned so it's using `latest`.
postgres 18 was released which uses a different directory structure
which in turn causes the DB to not work correctly. This causes our CI to
fail.

# Changes
Since we use postgres 16 in the prod infra I pinned it to this version
instead of 17 (which was the last working version).

## How to test
CI should work again