Migrate `CowProtocolToken` to alloy (#3808)
# Description
Migrates `CowProtocolToken` bindings from `ethcontract` to `alloy`.

Additionally this pins the `postgres` version to 16 since the release of
postgres 18 causes our CI to fail (due to a different folder structure).
Because the prod deployment uses 16 it makes the most sense to use in
our testing as well.

# Changes
- [x] removed old bindings
- [x] added alloy bindings
- [x] fixed resulting type errors
- [x] removed tests asserting that we have an address for all networks

## How to test
existing e2e tests