Make ABI files readable (#3815)
# Description
I'm still playing with the idea of reducing the ABI files to the subset
of types, functions, and events that we actually use in the code base.
That should cut down the compile time of the `contracts` crate quite a
lot.
There are quite a few ABI files that expose >30 functions of which we
only use 2 or 3. Since we still generate code for these 27 unused
functions this is just a lot of time wasted. (e.g. take a look at
[BalancerV2ComposableStablePool.json](https://github.com/cowprotocol/services/compare/format-abi-files?expand=1#diff-76cc4a1d98b9b1a835fe83e09a890ca42ebf19a1f94d73ede5701ae055e88a48))
So after this PR is merged it would be a lot easier to verify what
things we actually remove from the ABI files to speed up the compile
time (if we ever do that).

But even if we decide not to trim the files having them formatted into
multiple lines still helps with readability and also makes it easier for
text editors to handle these files.

# Changes
formatted all files with:
```fish
for f in *.json
        if test -L "$f"
            echo "Skipping symlink: $f"
            continue
        end
        jq . "$f" > "$f.tmp" && mv "$f.tmp" "$f"
    end
```

## How to test
look at the files, I guess. 😅