Replace `#[allow]` with `#[expect]` (#3825)
# Description
Rust
[`1.81`](https://blog.rust-lang.org/2024/09/05/Rust-1.81.0/#expect-lint)
introduced `#[expect]` as an alternative to `#[allow]`.
`#[allow]` always allows the given clippy lint to happen in the
annotated code regardless of whether it's actually necessary. A good
example is `#[allow(dead_code)]`. It's useful during development when
you don't have all the code hooked up yet. But once the code is hooked
up you might forget to remove the annotation. Once the code becomes
unused AGAIN (e.g. due to a refactor) you ideally would like to get
warned again.
This can be achieved with `#[expect]`. It also allows the provided
clippy lint to happen in the annotated code but it also complains if
clippy actually didn't complain about it while compiling.

So far my impression is `#[expect]` is strictly more useful than
`#[allow]` except for a few cases:
* macros - some uses of the code generated by the macro might trigger
`clippy` while others might not
* member variables that are only there for the side effects of `Drop` -
we don't care whether or not the field gets used outside of the `Drop`
logic
* debug - IIRC dead code analysis does not consider logging; as long as
we print the info we desire we don't care if the variable is used for
anything else

As a follow up it's also possible to enable a `clippy` lint for this but
this will only work once we move away from bindings generated with
`ethcontract-rs`.

# Changes
- replaced most occurrences of `#[allow]` with `#[expect]` (except where
it didn't make sense)
- deleted some outdated annotations that were caught with this change
- addressed some `dead_code` annotations

## How to test
compiler