Migrate CoW AMM-related SCs to alloy (#3813)
# Description
Migrates all the CoW AMM-related SCs to alloy with some caveats. The
`GPv2Order.Data` has different representation for its type hash and data
hash:
https://github.com/cowprotocol/contracts/blob/19972cd8fb3f8663846f772190926f36af068a33/src/contracts/libraries/GPv2Order.sol#L9-L48
Basically, for `kind`, `sellTokenBalance`, and `buyTokenBalance` fields,
the `byte32` type is used for the `data hash`, while for the `type
hash`, these fields are represented as strings, which was done for human
readability. Unfortunately, there is no easy way to manipulate ABI JSON
in a way that it starts using one struct for type hash and another for
data hash, since the full hash struct is just a combination of type and
data hashes, so a helper `GPv2OrderEip712` extension was introduced with
new functions to calculate the `GPv2Order.Data` hash.

## How to test
Existing tests.

## Further implementation
Once `GPv2Settlement` SC is migrated to alloy, it makes sense to drop
the `GPv2Order` definition from the cow-amm ABI JSON and switch to
`GPv2Settlement::GPv2Order`