Only fetch `auctionStartBlock` instead of the entire auction (#3847)
# Description
Currently we need to know the block at which we started the auction. We
do this by fetching the HUGE JSON that contains ALL the data of the
auction only to get 1 number out.


# Changes
This PR introduces a new query that only fetches that one number.

## How to test
existing tests for correctness
ran a build that does both queries and compares them in prod mainnet,
reduces the time from 20m to 2ms.