Speedup settlement indexing (simplified) (#3840)
# Description
Second attempt at speeding up settlement indexing. See
https://github.com/cowprotocol/services/pull/3654 for the context.
This PR again tries to speed up the underlying DB query that takes a
long time during the indexing step. Last time rebuilding the entire
query completely to squeeze out peak performance broke the indexing
logic which was very painful to recover from.
To not repeat this mistake this PR is a lot more conservative and only
changes 1 thing.

# Changes
Instead of fetching all orders of the entire original auction plus their
quotes and fee policies we only fetch those for the orders that were
actually traded in the given settlement. So the underlying queries are
identical we just try to avoid fetching irrelevant data.
For reference the only place where the queried fee policies are used is
[here](https://github.com/cowprotocol/services/blob/f8598be39cce699a68e0d1af5602fad945f165a9/crates/autopilot/src/domain/settlement/trade/math.rs#L208-L218)
where we need to look up the policy of every traded order to figure out
how much we collected in fees. Because there is no need to compute the
fees for orders that haven't been traded in the auction we can simply
not fetch fee policies for those orders.

I decided to not add any filtering logic to
`surplus_capturing_jit_order_owners` or `prices` because the queries
should be trivial and should mostly save network bandwidth but not
really query execution time.

## How to test
existing tests for correctness
briefly ran the new code on prod in a shadow mode and it lowers the time
for the DB queries from ~200ms to ~15ms.