add merge_group trigger on pull-request workflow (#3854)
We decide to enable merge queue as a merge requirement for our PRs.
This change is a prerequisite to enable merge queue on GitHub. It is how
the merge queue effectively triggers the CI, so it should lead to all
`required` marked actions to be triggered.

Find related conversation on
[slack](https://cowservices.slack.com/archives/C0375NV72SC/p1761853395235399).

# Description
This trigger is required for enabling merge queue support on GitHub (see
[docs](https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue#triggering-merge-group-checks-with-github-actions))

# Changes
Added the `merge_group` trigger.