[TRIVIAL] Log gas price of tx submission (#3862)
# Description
Currently our transaction submission logs are not very good when it
comes to figuring out why a transaction was not mined in a timely
manner. We are already able to recover the calldata of a transaction to
look for reverts but if the issue is a too low gas price we currently
wouldn't know.

# Changes
Also logs the `max_fee_per_gas` and `priority_fee_per_gas` after we
submitted a transaction to the mempool.