Ensure CoinGecko API key is set (#3905)
# Description

We ran into an issue where CoinGecko was enabled as native price
estimator, but there the necessary env var with API key was not set. To
prevent this in the future we want to have the service not start
instead. Otherwise one gets weird behaviour when the service runs where
sometimes we get native prices when some solvers decide to quote them,
but not always and that leads to order that would otherwise be solved
never get created in the first place as one can't get a quote ("no
liquidity" error).