Use standard hooks trampoline address (#3904)
# Description

Last months it wasn't possible to compile and deploy to Linea using
emv=shanghai, so we had to use an older version which didn't produce the
same bytecode hence a different address. They upgraded the chain sync
then, so we redeployed the contract and can now use the standard
address.