Do not preprocess COW AMMs when not configured (#3925)
# Description
When the driver's COW AMM config is empty, the corresponding task still
tries to search for surplus capturing jit order owners among the AMMs
and fires a redundant warning log in that case. When the driver doesn't
expect any AMMs, this log doesn't make any sense. This PR fixes this.

# Changes

- [ ] Make the COW AMM cache optional based on the COW AMMs config.
- [ ] When None, return fast with no AMMs for every auction with no
redundant logs.

## How to test
Logs on chains where COW AMMs are not configured.

## Related Issues

Fixes #3890