persist current auction in background task (#3861)
# Description
Currently the autopilot assembles the contents of the auction and then
persists it in the DB by replacing the current auction with the new one
which also returns the ID for the new auction.
This is unfortunately pretty slow (~440ms on prod mainnet) and therefor
significantly increases the delay between seeing a new block and sending
the fully assembled auction to solvers.


# Changes
What we can do instead is to introduce a new DB query that just
increments and returns the auction id counter for the next auction.
Afterwards we spawn a background task that writes the fully assembled
auction to the DB and uploads it to S3.

The trade-off is that it's now possible that we fully run an auction
that we don't have persisted anywhere. But in practice this does not
happen (no instance in the last
[30d](https://aws-es.cow.fi/_dashboards/app/data-explorer/discover#?_a=(discover:(columns:!(log,log_level),isDirty:!f,sort:!()),metadata:(indexPattern:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',view:discover))&_g=(filters:!(),refreshInterval:(pause:!t,value:0),time:(from:now-30d,to:now))&_q=(filters:!(('$state':(store:appState),meta:(alias:!n,disabled:!f,index:'86e4a5a0-4e4b-11ef-85c5-3946a99ed1a7',key:kubernetes.container_image,negate:!t,params:(query:nginx),type:phrase),query:(match_phrase:(kubernetes.container_image:nginx)))),query:(language:kuery,query:'%22failed%20to%20replace%20current%20auction%22'))))
and this information is for debugging purposes anyway.

## How to test
e2e tests