feat(autopilot): reduce excessive looping of main loop (#3938)
# Description
Currently the `run_fovever` loop in autopilot will continue looping
without waiting as long as there is no auction to process and the block
phase is still early enough. This is ofc not efficient and makes it
difficult to view debug logs in the playground as they are spammed every
time a block is mined. Technically an event listener is likely to be
able to process events faster than a 50ms database scanning loop (as it
is now)

# Changes
- [x] Add postgresql notification on new order added to order table
- [x] Listen to notification and trigger the main loop if it 
- [x] Update playground logs verbosity configuration to make it easier
to track loop status. Now that the logs don't spam as a result of
infinite looping, these changes make sense!
- [x] Verify that the loop is being executed on block and on order in
the playground

## How to test
1. Run the playground as normal `docker compose -f
playground/docker-compose.fork.yml up --build`
* you may have to delete and recreate your existing playground with
`docker compose -f playground/docker-compose.fork.yml down --volumes`
2. When all services stabilize, run `playground/test_playground.sh`
3. Observe logs indicating that the order even was received