Return deadline block in the solver competition V2 API (#3948)
# Description
Adds the auction deadline block to the solver competition V2 API as
requested in #3947:
> It would be very useful to add the /settle deadline in the competition
endpoint, as this would allow everyone to easily compute rewards based
on the transaction hashes already included there (and the corresponding
blocks txs took place) and the deadline of the /settle call.
> 
> This would also allow to easily extend the circuit breaker
functionality to blacklist a solver that executes after the deadline.

## How to test
Adjusted existing tests.

## Related Issues

Fixes #3947