Limit maintenance time (#3951)
# Description
Recently there was an issue where indexing 1 particular settlement took
a very long time. This caused the whole protocol to slow down which in
turn reduced the throughput and increased the time to happy moo.
This PR introduces a timeout for the maintenance logic. Whenever we hit
this timeout we prefer running a not fully updated auction over stalling
the protocol further (e.g. some order might be included which has
already been settled).