Keep tokio console only in the playground (#3950)
# Description
It turned out that `tokio_unstable` introduces a memory leak when
creating spans. As @MartinquaXD noticed:
> Most of the tracing related memory was allocated while holding a tokio
mutex lock. When you check the tokio code you see that this only happens
when we build it with `tokio_unstable` which is only used for
`tokio-console` support.

Disabling the `tokio_unstable` helped to mitigate the memory leak issue
almost completely.

# Changes

- Remove `tokio_unstable` from the codebase.
- Keep it only in the playground (for now).

## Further implementation
Using the updated deployment CI job from #3954, the tokio-console can be
gated behind a feature in case we need to run it in the future. Will
implement it in a follow-up pr.

(cherry picked from commit 70b17eca136c774174b12f3f6b0cca2ae4b418f2)