Parallelize post processing DB operations further (#3955)
# Description
Post processing (i.e. storing a bunch of data in the DB) is on the
critical auction path. While it does not increase the latency between
building the auction and solvers receiving it it does increase the time
between getting all solutions and telling the winner to settle.


# Changes
The logic was already split into 2 chunks of concurrent tasks. This PR
just merges the chunks into one. The new post-processing time should be
`max(all_operations)` instead of `max(chunk1_ops) + max(chunk2_ops)`.

Also 1 of the DB queries used a transaction unnecessarily so I changed
that to use a regular connection.

## How to test
e2e tests