Better tx replacement logic for solution submission (#3941)
With the recent gnosis chain issues we saw a lot of issues with
replacing transactions. Additionally it was not very easy to understand
what went wrong due to insufficient logging. This PR addresses both
issues.

Logic:
* compute a fresh gas price before submitting the settlement
* always check if there is a pending tx we have to replace and pick the
gas price accordingly
* simple mempool specific logic is in the `infra` module
* pricing decisions happen in the `domain` module
* slightly bump the theoretical minimum of 12.5% gas price increase per
block to 30% to avoid numerical issues for tiny gas prices like on
gnosis chain

Logging:
* always log success / failure directly in the `submit()` function
* on error also try to fetch the current pending tx (in case our
submission failed because there was a higher priced pending tx in the
mempool)
* log many intermediate gas prices for understanding the final results
of non-trivial pricing decisions

Tested the logic in prod on gnosis chain where we struggled a lot with
replacing transactions correctly.