Use only scores computed by the autopilot (#3978)
# Description
Currently drivers are required to compute a score for their solution and
commit to it. If the scores computed by the driver are different from
what the autopilot computes it can result in situations where not enough
rewards are paid out from the solver accounting scripts.
Since the autopilot has to compute scores anyway we might as well only
use and store those instead of the possibly buggy driver computed
scores.

Addresses parts of https://github.com/cowprotocol/services/issues/3462.

# Changes
Given how pressing this situation has become this PR only applies a
bandage and foregoes a "proper" refactor of the involved code.
- removed `score` from the autopilot `/solve` response DTO
- removed driver reported score from the `autopilot` domain types
- always read the autopilot computed score when storing proposed
solutions in the DB
- removes `score` handling and assertions from `driver` tests

The ugly part is that we first build a `Solution` and only later have
the winner selection logic populate the `score`. This technically
exposes us to the problem where the score gets read before it has been
set. However, this does not happen in practice because ranking happens
immediately after receiving the solutions and only solutions where a
score could be computed get preserved. (see
[here](https://github.com/cowprotocol/services/pull/3978/changes#diff-d541742700ed647c8a6f817fd9672344ceeef30de024053b8f56146d49349cafR253-R277))

I also opted to still have the reference driver report the computed
score as there are quite a few driver integration tests that test subtle
scoring related logic. Losing those would be a shame and refactoring
those to work without reporting the score in the response is also a ton
of work.

This will be followed up with a PR refactoring the code such that the
score gets initialized during construction to avoid this 2 phase
initialization process.

## How to test
existing e2e tests